refactor(DataType): 重构所有 id 数据类型
将所有 Long 类型 id，修改为 Integer 类型

+++ src/main/java/io/github/talelin/latticy/service/UserIdentityService.java
@@ -80,5 +80,5 @@ public interface UserIdentityService extends IService<UserIdentityDO> {
      * @param password 新密码
      * @return 是否成功
      */
-    boolean changeUsernamePassword(Long userId, String username, String password);
+    boolean changeUsernamePassword(Integer userId, String username, String password);
 }
